+++ src/js/server/controllers/apply/dashboard-controller.ts
@@ -1,6 +1,6 @@
 import { RequestHandler } from 'express';
 
-import { UserRequest } from 'js/server/apply/router';
+import { UserRequest } from 'js/server/routes/apply-router';
 import * as utils from 'js/server/utils.js';
 import { Hacker, TeamMember, HackerInstance } from 'js/server/models';
 import * as statuses from 'js/shared/status-constants';
