+++ src/js/server/routes/index.ts
@@ -1,0 +1,1 @@
+require('router');
